+++ sdk/cognitivelanguage/azure-ai-language-questionanswering/azure/ai/language/questionanswering/projects/_patch.py
@@ -86,5 +86,5 @@ class QuestionAnsweringProjectsClient(QuestionAnsweringProjectsClientGenerated):
 
 
 def patch_sdk():
-    curr_package = importlib.import_module("azure.ai.language.questionanswering.projects")
-    curr_package.QuestionAnsweringProjectsClient = QuestionAnsweringProjectsClient
+    current_package = importlib.import_module("azure.ai.language.questionanswering.projects")
+    current_package.QuestionAnsweringProjectsClient = QuestionAnsweringProjectsClient