+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/aio/_document_model_administration_client_async.py
@@ -28,8 +28,8 @@ from .._models import (
     ModelBuildMode,
     DocumentModelDetails,
     DocumentModelSummary,
-    DocumentModelOperationDetails,
-    DocumentModelOperationSummary,
+    OperationDetails,
+    OperationSummary,
     ResourceDetails,
     TargetAuthorization,
 )
@@ -430,15 +430,15 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         return DocumentModelDetails._from_generated(response)
 
     @distributed_trace
-    def list_operations(self, **kwargs: Any) -> AsyncItemPaged[DocumentModelOperationSummary]:
+    def list_operations(self, **kwargs: Any) -> AsyncItemPaged[OperationSummary]:
         """List information for each document model operation.
 
         Lists all document model operations associated with the Form Recognizer resource.
         Note that operation information only persists for 24 hours. If the operation was successful,
         the document model can be accessed using the :func:`~get_document_model` or :func:`~list_document_models` APIs.
 
-        :return: A pageable of DocumentModelOperationSummary.
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[DocumentModelOperationSummary]
+        :return: A pageable of OperationSummary.
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[OperationSummary]
         :raises ~azure.core.exceptions.HttpResponseError:
 
         .. admonition:: Example:
@@ -454,13 +454,13 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         return self._client.get_operations(  # type: ignore
             cls=kwargs.pop(
                 "cls",
-                lambda objs: [DocumentModelOperationSummary._from_generated(x) for x in objs],
+                lambda objs: [OperationSummary._from_generated(x) for x in objs],
             ),
             **kwargs
         )
 
     @distributed_trace_async
-    async def get_operation(self, operation_id: str, **kwargs: Any) -> DocumentModelOperationDetails:
+    async def get_operation(self, operation_id: str, **kwargs: Any) -> OperationDetails:
         """Get a model operation by its ID.
 
         Get a model operation associated with the Form Recognizer resource.
@@ -468,8 +468,8 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         the model can be accessed using the :func:`~get_document_model` or :func:`~list_document_models` APIs.
 
         :param str operation_id: The operation ID.
-        :return: DocumentModelOperationDetails
-        :rtype: ~azure.ai.formrecognizer.DocumentModelOperationDetails
+        :return: OperationDetails
+        :rtype: ~azure.ai.formrecognizer.OperationDetails
         :raises ~azure.core.exceptions.HttpResponseError:
 
         .. admonition:: Example:
@@ -485,7 +485,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         if not operation_id:
             raise ValueError("'operation_id' cannot be None or empty.")
 
-        return DocumentModelOperationDetails._from_generated(
+        return OperationDetails._from_generated(
             await self._client.get_operation(operation_id, **kwargs),
             api_version=self._api_version,
         )