+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_dmac_mgmt_async.py
@@ -14,7 +14,7 @@ from azure.core.exceptions import ResourceNotFoundError, ClientAuthenticationErr
 from azure.ai.formrecognizer import (
     DocumentModelAdministrationClient,
     DocumentAnalysisApiVersion,
-    DocumentModelOperationDetails
+    OperationDetails
 )
 from azure.ai.formrecognizer.aio import DocumentModelAdministrationClient
 from preparers import FormRecognizerPreparer
@@ -84,8 +84,8 @@ class TestManagementAsync(AsyncFormRecognizerTest):
         async with client:
             info = await client.get_resource_details()
 
-        assert info.document_model_limit
-        assert info.document_model_count
+        assert info.custom_document_models.limit
+        assert info.custom_document_models.count
 
     @FormRecognizerPreparer()
     @DocumentModelAdministrationClientPreparer()
@@ -161,7 +161,7 @@ class TestManagementAsync(AsyncFormRecognizerTest):
                 op = await client.get_operation(successful_op.operation_id)
                 # test to/from dict
                 op_dict = op.to_dict()
-                op = DocumentModelOperationDetails.from_dict(op_dict)
+                op = OperationDetails.from_dict(op_dict)
                 assert op.error is None
                 model = op.result
                 assert model.model_id
@@ -181,7 +181,7 @@ class TestManagementAsync(AsyncFormRecognizerTest):
                 op = await client.get_operation(failed_op.operation_id)
                 # test to/from dict
                 op_dict = op.to_dict()
-                op = DocumentModelOperationDetails.from_dict(op_dict)
+                op = OperationDetails.from_dict(op_dict)
 
                 error = op.error
                 assert op.result is None